Support for Linux/x86-64, not much different from Linux/x86.

+++ tests/test_pipe_bytecount.c
@@ -112,7 +112,7 @@ int main(void) {
         if (retval == this_transfer)
             write(1, "#", 1);
         else {
-            printf("[%d]", retval);
+            printf("[%ld]", (long) retval);
             fflush(stdout);
         }
         total_rcvd += retval;
